refactor: move utils into separate dir
