+++ openassessment/xblock/staff_area_mixin.py
@@ -14,10 +14,10 @@ from submissions.errors import SubmissionNotFoundError
 from openassessment.assessment.errors import PeerAssessmentInternalError
 from openassessment.fileupload.api import delete_shared_files_for_team, remove_file
 from openassessment.workflow.errors import AssessmentWorkflowError, AssessmentWorkflowInternalError
-from openassessment.xblock.data_conversion import create_submission_dict
-from openassessment.xblock.resolve_dates import DISTANT_FUTURE, DISTANT_PAST
+from openassessment.xblock.utils.data_conversion import create_submission_dict
+from openassessment.xblock.utils.resolve_dates import DISTANT_FUTURE, DISTANT_PAST
 
-from .user_data import get_user_preferences
+from .utils.user_data import get_user_preferences
 
 logger = logging.getLogger(__name__)  # pylint: disable=invalid-name
