+++ openassessment/xblock/test/test_defaults.py
@@ -5,8 +5,8 @@ Basic tests for data defaults.
 import ddt
 from django.test import TestCase
 
-from openassessment.xblock import defaults
-from openassessment.xblock.validation import _is_valid_assessment_sequence as is_valid_assessment_sequence
+from openassessment.xblock.utils import defaults
+from openassessment.xblock.utils.validation import _is_valid_assessment_sequence as is_valid_assessment_sequence
 
 
 @ddt.ddt
