+++ openassessment/xblock/test/test_leaderboard.py
@@ -16,7 +16,7 @@ import boto3
 from moto import mock_s3
 from submissions import api as sub_api
 from openassessment.fileupload import api
-from openassessment.xblock.data_conversion import create_submission_dict, prepare_submission_for_serialization
+from openassessment.xblock.utils.data_conversion import create_submission_dict, prepare_submission_for_serialization
 
 from .base import XBlockHandlerTransactionTestCase, scenario
