+++ openassessment/xblock/test/test_resolve_dates.py
@@ -10,8 +10,8 @@ import pytz
 from django.test import TestCase
 
 from workbench.runtime import WorkBenchUserService
-from openassessment.xblock.resolve_dates import DISTANT_FUTURE, DISTANT_PAST, resolve_dates
-from openassessment.xblock.user_data import get_user_preferences
+from openassessment.xblock.utils.resolve_dates import DISTANT_FUTURE, DISTANT_PAST, resolve_dates
+from openassessment.xblock.utils.user_data import get_user_preferences
 
 
 def STUB_I18N(x):
