+++ openassessment/xblock/test/test_validation.py
@@ -11,7 +11,7 @@ import pytz
 from django.test import TestCase
 
 from openassessment.xblock.openassessmentblock import OpenAssessmentBlock
-from openassessment.xblock.validation import (validate_assessment_examples, validate_assessments, validate_dates,
+from openassessment.xblock.utils.validation import (validate_assessment_examples, validate_assessments, validate_dates,
                                               validate_rubric, validate_submission, validator)
